# swagger.yaml
swagger: '2.0'
info:
  title: MCP Streamable HTTP Server
  description: >-
    Multi-agent post generator with MCP streamable protocol support.
    Compatible with Microsoft Copilot Studio.
  version: 1.0.0
host: langfuseappservice.azurewebsites.net
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /mcp/tools/call:
    post:
      summary: Call MCP Tool
      x-ms-agentic-protocol: mcp-streamable-1.0
      description: Call MCP tools using FastMCP with automatic streaming support
      operationId: CallMCPTool
      consumes:
        - application/json
      produces:
        - application/json
        - text/event-stream
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - method
              - params
            properties:
              method:
                type: string
                description: Method name (tools/call)
                default: tools/call
              params:
                type: object
                required:
                  - name
                  - arguments
                properties:
                  name:
                    type: string
                    description: Tool name
                    enum:
                      - generate_linkedin_post
                      - generate_linkedin_post_streaming
                  arguments:
                    type: object
                    properties:
                      topic:
                        type: string
                        description: Topic for the LinkedIn post
                      user_id:
                        type: string
                        description: Optional user identifier
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
...
security: []
...
securityDefinitions: {}